Copy org.freedesktop.Sdk.Extension.gcc7 libs into app
Both libgcc_s.so.1 and libstdc++.so.6 have evolved (backwards-compatibly) from
GCC 6 (whose libraries are included in the org.gnome.Platform/3.24 runtime) to 7
(<https://gcc.gnu.org/onlinedocs/gcc-7.3.0/libstdc++/manual/manual/abi.html>),
so there is a chance that any of the LO binaries might use new symbols from the
updated libs.

The wildcard when copying libstdc++.so.6* is so that
/app/lib/libstdc++.so.6.0.24-gdb.py will be included, in case that's useful for
debugging.

The build-time LD_LIBRARY_PATH is in case any of the binaries run during the
build (e.g., tools and test code) might use new symbols from the updated libs.