Revert building on arm again, lack of org.freedesktop.Sdk.Extension.openjdk9
<https://github.com/flathub/org.freedesktop.Sdk.Extension.openjdk9/commit/
dca4e3ab608ab430b0086359343d2994600979dc> "Initial version":  "This doesn't
build for arm, because the debian openjdk for arm which is currently used to
bootstrap is a 'client' jvm variant, and the build needs a 'server' variant to
succeed."

So disable arm again for now, until either the SDK extension is available, or a
mechanism is provided in flatpak-manifest to exclude SDK extensions for specific
arches.

This reverts commit 6ab63b844c3a14878763ddba3efe3888e5afa457 "allow building on
arm".