rename sub-app .desktop files so they are strict sub-IDs
The org.libreoffice.LibreOffice-foo names are not actually sub-IDs of
org.libreoffice.LibreOffice which means that Flatpak will not export the
additional components in the appstream. Change the - to . so that the
sub-apps are correctly within the org.libreoffice.LibreOffice namespace
rather than a mere prefix match. Set the X-Flatpak-RenamedFrom list
in the .desktop files to include both the original upstream names
and the - delimited old IDs that we used previous in the Flatpak.

https://github.com/flathub/org.libreoffice.LibreOffice/issues/36